requirements.txt
flask
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
pypdf
faiss-cpu
sentence-transformers
python-dotenv
openai
